requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
onnxruntime
python-multipart
slowapi
python-dotenv
gunicorn
redis
mangum
